Move locator non-gui classes to core
Allows reuse in projects which don't build QGIS gui, e.g.
QField
