+++ src/core/locator/qgslocator.cpp
