+++ src/core/locator/qgslocatorfilter.cpp
